pyproject.toml
[tool.poetry]
name = "recruitment-platform"
version = "0.1.0"
description = ""
authors = ["Abhishek Saini <abhishek.saini@tyroo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
